requirements.txt
APScheduler==3.10.4
blinker==1.7.0
bson==0.5.10
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.4
dnspython==2.5.0
Flask==3.0.1
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
rsa==4.9
six==1.16.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
zope.interface==6.1
